+++ e2e/mobile/specs/delegate/delegateOSMO.spec.ts
@@ -1,0 +1,4 @@
+import { runDelegateTest } from "./delegate";
+
+const delegation = new Delegate(Account.OSMO_1, "0.0001", "Ledger by Figment");
+runDelegateTest(delegation, ["B2CQA-3022"]);