+++ libs/device-react/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/device-react
 
-## 0.2.30-nightly.0
+## 0.2.30
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @ledgerhq/device-core@0.4.24-nightly.0
+  - @ledgerhq/device-core@0.4.24
+
+## 0.2.30-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/device-core@0.4.24-next.0
 
 ## 0.2.29