+++ libs/ledgerjs/packages/swift-bridge-hw-app-eth/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/swift-bridge-hw-app-eth
 
-## 1.2.31-nightly.0
+## 1.2.31
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @ledgerhq/hw-app-eth@6.45.4-nightly.0
+  - @ledgerhq/hw-app-eth@6.45.4
+
+## 1.2.31-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.4-next.0
 
 ## 1.2.30